backend/requirements.txt
annotated-types==0.7.0
asyncpg
bcrypt==3.2.2
fastapi==0.116.1
httpcore==1.0.9
httpx==0.28.1
jose==1.0.0
numpy==2.3.2
openpyxl==3.1.5
passlib==1.7.4
psycopg2-binary
pyasn1==0.6.1
pycparser==2.22
pydantic
pydantic-settings
pydantic_core
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.47.3
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.5.0
uvicorn==0.35.0